modificar Notra

+++ EjercicioNotas/src/com/krakedev/entidades/Nota.java
@@ -2,7 +2,7 @@ package com.krakedev.entidades;
 
 public class Nota {
 	private Materia materia;
-	private int calificacion;
+	private double calificacion;
 	
 	//metodos
 	public void mostrar() {
@@ -16,10 +16,10 @@ public class Nota {
 	public void setMateria(Materia materia) {
 		this.materia = materia;
 	}
-	public int getCalificacion() {
+	public double getCalificacion() {
 		return calificacion;
 	}
-	public void setCalificacion(int calificacion) {
+	public void setCalificacion(double calificacion) {
 		this.calificacion = calificacion;
 	}
